+++ web/src/components/UsageHeatMap.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useRef, useState } from "react";
-import { useAppSelector } from "../store";
-import { locationService, userService } from "../services";
+import { useLocationStore, useMemoStore, useUserStore } from "../store/module";
 import { getMemoStats } from "../helpers/api";
 import { DAILY_TIMESTAMP } from "../helpers/consts";
 import * as utils from "../helpers/utils";
@@ -28,19 +27,21 @@ interface DailyUsageStat {
 }
 
 const UsageHeatMap = () => {
+  const locationStore = useLocationStore();
+  const userStore = useUserStore();
+  const memoStore = useMemoStore();
   const todayTimeStamp = utils.getDateStampByDate(Date.now());
   const todayDay = new Date(todayTimeStamp).getDay() + 1;
   const nullCell = new Array(7 - todayDay).fill(0);
   const usedDaysAmount = (tableConfig.width - 1) * tableConfig.height + todayDay;
   const beginDayTimestamp = todayTimeStamp - usedDaysAmount * DAILY_TIMESTAMP;
-
-  const { memos } = useAppSelector((state) => state.memo);
+  const memos = memoStore.state.memos;
   const [allStat, setAllStat] = useState<DailyUsageStat[]>(getInitialUsageStat(usedDaysAmount, beginDayTimestamp));
   const [currentStat, setCurrentStat] = useState<DailyUsageStat | null>(null);
   const containerElRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    getMemoStats(userService.getCurrentUserId())
+    getMemoStats(userStore.getCurrentUserId())
       .then(({ data: { data } }) => {
         const newStat: DailyUsageStat[] = getInitialUsageStat(usedDaysAmount, beginDayTimestamp);
         for (const record of data) {
@@ -84,11 +85,11 @@ const UsageHeatMap = () => {
   }, []);
 
   const handleUsageStatItemClick = useCallback((item: DailyUsageStat) => {
-    if (locationService.getState().query?.duration?.from === item.timestamp) {
-      locationService.setFromAndToQuery();
+    if (locationStore.getState().query?.duration?.from === item.timestamp) {
+      locationStore.setFromAndToQuery();
       setCurrentStat(null);
     } else if (item.count > 0) {
-      locationService.setFromAndToQuery(item.timestamp, item.timestamp + DAILY_TIMESTAMP);
+      locationStore.setFromAndToQuery(item.timestamp, item.timestamp + DAILY_TIMESTAMP);
       setCurrentStat(item);
     }
   }, []);
