--- web/src/services/README.md
@@ -1,6 +1,0 @@
-# Services
-
-What should service do?
-
-- request data api and throw error;
-- dispatch state actions;
