--- web/src/services/index.ts
@@ -1,8 +1,0 @@
-import globalService from "./globalService";
-import locationService from "./locationService";
-import memoService from "./memoService";
-import shortcutService from "./shortcutService";
-import userService from "./userService";
-import resourceService from "./resourceService";
-
-export { globalService, locationService, memoService, shortcutService, userService, resourceService };
