+++ web/src/store/module/index.ts
@@ -1,1 +1,7 @@
 export * from "./editor";
+export * from "./global";
+export * from "./location";
+export * from "./memo";
+export * from "./resource";
+export * from "./shortcut";
+export * from "./user";
